ai_web_scraper_agent/requirements.txt
beautifulsoup4
requests
playwright
streamlit
openai  # Optional if you use any free model via OpenRouter
python-dotenv
